# frontend/openapi.json
{
  "components": {
    "schemas": {
      "ChatRequest": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          }
        },
        "required": [
          "message"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
      "ChatResponse": {
        "properties": {
          "errors": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Errors",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "default": {},
            "title": "Metadata",
            "type": "object"
          },
          "processing_time_ms": {
            "title": "Processing Time Ms",
            "type": "number"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "string"
          },
          "tools_used": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Tools Used",
            "type": "array"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "success",
          "message",
          "session_id",
          "user_id",
          "timestamp",
          "processing_time_ms"
        ],
        "title": "ChatResponse",
        "type": "object"
      },
...
      "CreateSessionRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "CreateSessionRequest",
        "type": "object"
      },
      "CreateSessionResponse": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "errors": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Errors",
            "type": "array"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "success",
          "session_id",
          "title",
          "created_at"
        ],
        "title": "CreateSessionResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SessionHistoryResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "errors": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Errors",
            "type": "array"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/SessionMessage"
            },
            "title": "Messages",
            "type": "array"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "session_id",
          "messages",
          "count"
        ],
        "title": "SessionHistoryResponse",
        "type": "object"
      },
      "SessionListItem": {
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "last_message_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Message At"
          },
          "message_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Count"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "title"
        ],
        "title": "SessionListItem",
        "type": "object"
      },
      "SessionListResponse": {
        "properties": {
          "errors": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Errors",
            "type": "array"
          },
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "sessions": {
            "items": {
              "$ref": "#/components/schemas/SessionListItem"
            },
            "title": "Sessions",
            "type": "array"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "success",
          "sessions",
          "total",
          "page",
          "limit",
          "has_more"
        ],
        "title": "SessionListResponse",
        "type": "object"
      },
      "SessionMessage": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "required": [
          "role",
          "content"
        ],
        "title": "SessionMessage",
        "type": "object"
      },
...
      "UpdateSessionTitleRequest": {
        "properties": {
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "UpdateSessionTitleRequest",
        "type": "object"
      },
      "UpdateSessionTitleResponse": {
        "properties": {
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "success",
          "session_id",
          "title"
        ],
        "title": "UpdateSessionTitleResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "JWT": {
        "description": "JWT authentication using Bearer token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AI-powered agent hub with multi-provider LLM support",
    "title": "AgentHub API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/chat/message": {
      "post": {
        "description": "Send a message to the AI agent.\n\nRequires authentication via JWT Bearer token.\n\n- **message**: The user's message/question\n- **session_id**: Optional session ID. If not provided, creates a new session\n\nReturns the agent's response with metadata.",
        "operationId": "send_message_api_v1_chat_message_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Send Message",
        "tags": [
          "chat"
        ]
      }
    },
    "/api/v1/chat/sessions": {
      "get": {
        "description": "List all chat sessions for a user.\n\nRequires authentication via JWT Bearer token.\n\n- **page**: Page number (0-based, default: 0)\n- **limit**: Number of sessions per page (default: 10)\n\nReturns paginated list of user's sessions.",
        "operationId": "list_user_sessions_api_v1_chat_sessions_get",
        "parameters": [
          {
            "description": "Page number (0-based)",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Page number (0-based)",
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Number of sessions per page",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "description": "Number of sessions per page",
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "List User Sessions",
        "tags": [
          "chat"
        ]
      },
      "post": {
        "description": "Create a new chat session.\n\nRequires authentication via JWT Bearer token.\n\n- **title**: Optional title for the session\n\nReturns the new session details.",
        "operationId": "create_session_api_v1_chat_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Create Session",
        "tags": [
          "chat"
        ]
      }
    },
    "/api/v1/chat/sessions/{session_id}/messages": {
      "get": {
        "description": "Get the message history for a specific session.\n\nRequires authentication via JWT Bearer token.\n\n- **session_id**: ID of the session\n- **limit**: Maximum number of messages to return (default: 50)\n\nReturns the session's message history.",
        "operationId": "get_session_history_api_v1_chat_sessions__session_id__messages_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "description": "Maximum number of messages to return",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "default": 50,
              "description": "Maximum number of messages to return",
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Get Session History",
        "tags": [
          "chat"
        ]
      }
    },
    "/api/v1/chat/sessions/{session_id}/title": {
      "put": {
        "description": "Update the title of a chat session.\n\nRequires authentication via JWT Bearer token.\n\n- **session_id**: ID of the session to update\n- **title**: New title for the session\n\nReturns the updated session details.",
        "operationId": "update_session_title_api_v1_chat_sessions__session_id__title_put",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSessionTitleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateSessionTitleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Update Session Title",
        "tags": [
          "chat"
        ]
      }
    },
...
  }
}